+++ sw/src/hall.c
@@ -182,9 +182,9 @@ void TIM4_IRQHandler(void)
 
 static uint8_t hall_state_update(hall_state_t * const state)
 {
-  const uint8_t h1 = ((HALL_SENSOR_H1_PORT->IDR & HALL_SENSOR_H1_PIN) == HALL_SENSOR_H1_PIN);
-  const uint8_t h2 = ((HALL_SENSOR_H2_PORT->IDR & HALL_SENSOR_H2_PIN) == HALL_SENSOR_H2_PIN);
-  const uint8_t h3 = ((HALL_SENSOR_H3_PORT->IDR & HALL_SENSOR_H3_PIN) == HALL_SENSOR_H3_PIN);
+  const uint8_t h1 = (uint8_t) ((HALL_SENSOR_H1_PORT->IDR & HALL_SENSOR_H1_PIN) == HALL_SENSOR_H1_PIN);
+  const uint8_t h2 = (uint8_t) ((HALL_SENSOR_H2_PORT->IDR & HALL_SENSOR_H2_PIN) == HALL_SENSOR_H2_PIN);
+  const uint8_t h3 = (uint8_t) ((HALL_SENSOR_H3_PORT->IDR & HALL_SENSOR_H3_PIN) == HALL_SENSOR_H3_PIN);
 
   state->previous = state->current;
   state->current = ((h3 << 2) | (h2 << 1) | (h1));
